requirements.txt
httpx==0.24.0
python-telegram-bot[callback-data]==20.3
requests
yt-dlp==2023.3.4
filetype==1.2.0
cachetools~=5.3.0